requirements.txt
fastapi==0.103.1
uvicorn==0.23.2 
python-dotenv==1.0.0
supabase==1.0.4
pandas==2.0.1